+++ src/libImaging/BitDecode.c
@@ -20,7 +20,7 @@
 
 
 int
-ImagingBitDecode(Imaging im, ImagingCodecState state, UINT8* buf, int bytes)
+ImagingBitDecode(Imaging im, ImagingCodecState state, UINT8* buf, Py_ssize_t bytes)
 {
     BITSTATE* bitstate = state->context;
     UINT8* ptr;
