+++ src/libImaging/RawDecode.c
@@ -20,7 +20,7 @@
 
 
 int
-ImagingRawDecode(Imaging im, ImagingCodecState state, UINT8* buf, int bytes)
+ImagingRawDecode(Imaging im, ImagingCodecState state, UINT8* buf, Py_ssize_t bytes)
 {
     enum { LINE = 1, SKIP };
     RAWSTATE* rawstate = state->context;
